あ

+++ lib/pbmenv.rb
@@ -26,16 +26,30 @@ module Pbmenv
     # curl -L https://github.com/splaplapla/procon_bypass_man/archive/refs/tags/v0.1.6.tar.gz | tar xvz
     # git clone https://github.com/splaplapla/procon_bypass_man.git -b v0.1.6 procon_bypass_man-0.1.6
     # git clone https://github.com/splaplapla/procon_bypass_man.git procon_bypass_man-0.1.6
-    shell = <<~SHELL
-      git clone https://github.com/splaplapla/procon_bypass_man.git procon_bypass_man-0.1.6
+    if ENV["DEBUG_INSTALL"]
+      shell = <<~SHELL
+        git clone https://github.com/splaplapla/procon_bypass_man.git procon_bypass_man-#{version}
+      SHELL
+    else
+      shell = <<~SHELL
+        curl -L https://github.com/splaplapla/procon_bypass_man/archive/refs/tags/v#{version}.tar.gz | tar xvz
+      SHELL
+    end
+    system(shell)
+    unless File.exists?("procon_bypass_man-#{version}/project_template")
+      raise "This version is not support by pbmenv"
+    end
+
+    system <<~SHELL
       mkdir -p #{PBM_DIR}/v#{version}
       cp -r procon_bypass_man-#{version}/project_template/* #{PBM_DIR}/v#{version}/
     SHELL
-    system shell
     use version
   rescue => e
     system "rm -rf #{PBM_DIR}/v#{version}"
     raise
+  ensure
+    system "rm -rf #{PBM_DIR}/procon_bypass_man-#{version}"
   end
 
   def self.uninstall(version)
@@ -44,10 +58,12 @@ module Pbmenv
 
   def self.use(version)
     unless File.exists?("/usr/share/pbm/v#{version}")
-      false
+      return false
     end
 
-    system "unlink #{PBM_DIR}/current"
+    if File.exists?("#{PBM_DIR}/current")
+      system "unlink #{PBM_DIR}/current"
+    end
     system "ln -s #{PBM_DIR}/current #{PBM_DIR}/v#{version}"
   end
 end
